Split ByGenericTypeRecord into two
